moved detect nan to training loop, remove flag for print
